+++ src/service/submission/submission.hook.service.ts
@@ -59,9 +59,9 @@ export class SubmissionHookService {
         return {
           field: submissionField.field.id,
           slug: submissionField.field.slug || null,
-          default_value: submissionField.field.value,
+          default_value: submissionField.field.defaultValue,
           content: submissionField.content,
-	}
+        }
       }),
     }